+++ ui/components/app/snaps/snap-ui-skeleton/index.ts
